requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-cli==0.1
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
option==2.1.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
PyAudio==0.2.14
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
replicate==0.22.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SpeechRecognition==3.10.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
tornado==6.4
tqdm==4.66.1
transformers==4.36.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
